begins reorganizing components
