+++ packages/cms/src/components/fields/Button.css
@@ -1,4 +1,4 @@
-@import "../css/mixins.css";
+@import "../../css/mixins.css";
 
 /* buttons */
 .cms-button {
