+++ packages/cms/src/components/fields/ButtonGroup.css
@@ -1,4 +1,4 @@
-@import "../css/mixins.css";
+@import "../../css/mixins.css";
 
 .cms-button-group {
   /* layout */
