+++ packages/cms/src/components/fields/Select.css
@@ -1,4 +1,4 @@
-@import "../css/mixins.css";
+@import "../../css/mixins.css";
 
 /* wrapper */
 .cms-select-label {
