+++ packages/cms/src/index.js
@@ -2,5 +2,5 @@ export {default as Builder} from "./Builder.jsx";
 
 // Components
 export {default as Profile} from "./components/Profile.jsx";
-export {default as Section} from "./components/Section.jsx";
-export {default as Search} from "./components/Search/Search.jsx";
+export {default as Section} from "./components/sections/Section.jsx";
+export {default as Search} from "./components/fields/Search.jsx";
